requirements.txt
google-generativeai>=0.3.0
flask>=2.0.0
flask-wtf>=1.0.0
flask-sqlalchemy>=3.0.0
matplotlib>=3.5.0
pandas>=1.3.0
openpyxl>=3.0.0
pytest>=7.0.0
pytest-flask>=1.2.0
validators